parse: Add and partially document the parser.
The parser implements left-associative function application, but
no lambdas.  The idea is to get some kind of AST and in and out.

This commit is big, but it leaves out proper error hadling - we
just die on parse errors.  We don't even use DIE_IF so the coverage
is far from 100%