type: Hackishly handling a recursive type.
The feature we need here is to track the bindings as we build
the type tree.  We do not need the recursive unify.

But the problem here is that the unparse_type function is already
recursive (even though I suppose it doesn't have to be).  So
we hack around this with a brutal depth=1 rule.