type: unparser tracks the stack.
Unparser can now reasonably deep type trees, including recursive ones.  At
least we have got rid of the dumb depth=1 limit.  Instead we use an explicit
stack (as well the C stack!).  So we can cut of recursion by scanning back
through that stack.  We impose a fixed-and-fatal limit on the stack size.