type: unify_nonrecursive_functions_shallowly
We have a unify function, which can even operate on function, but doesn't do a
proper job of unifying the components -- it only remasters the function type
itself.

That is the unify function is nonrecrsuive.  But here "nonrecursive" actually
means the functions we are working on are nonrecursive.

The next step will still be with nonrecursive test functions, but ones design
to require a proper recursive unify() funciton.