type: out-of-order repaired.
We handle the out-of-order case differently: now we enforce the rule of always
unifying into the prior type.  If the newer type turns out to be a function, we
*copy* into the prior before twiddling the union-find.

Modulo name-twiddling this change has no effect.  However it produces surprises
with auto-naming.

Initially we have two separate types

        A = ... free type var ...
        B = (Y Br)

Then we unify them, since A is prior to B, we unify B into A and the result
is called:

        A = (Y Br)