test: unparse_type does not touch the types.
That is is the unparser works with const pointers.  This is a non-trivial
change because it means the unparser is not allowed to call masterse(), which
is a mutating function.  The unparser *could* just chase the links, but
instead, after solving types, I do a pre-scan and masterise in build_type_tree.

There is still an issue that the only way to know if two types are the same is
to try and unify them.  (Which at the moment is useless, since unification
always succeeds.)