lambda: parsing and typing trivial lambdas
The user can introduce lambdas, both with named and unnamed params.  In BOTH
cases a freevar node is emitted after the body, and then a lambda nodes is
emitted.

This commit also adds a coerce_lambda to the typer, which joins coerce callee
in defining function types.

There are tests for some simple lambdas, as well as for each parse errors that
can occur.