DBG: Tests properly ignore "DBG: " prefix
Previously, tests would scan stderr and check that every line began with "DBG:
".  This would only happen if we were not expecting errors -- which means
adding DBG output would break bad-path tests.

Now there is a pre-filter step to strip the DBG lines (and empty ones).  On the
good-path, we check that the output of the filter is empty.  On bad-path tests
the filtered output is passed on for further processing.